# mobile/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support-tickets]
      summary: List tickets
      parameters:
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/TicketStatus"
        - in: query
          name: priority
          required: false
          schema:
            $ref: "#/components/schemas/TicketPriority"
        - in: query
          name: assigneeId
          required: false
          schema:
            type: string
        - in: query
          name: search
          required: false
          schema:
            type: string
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [updated_desc, updated_asc, created_desc, priority_desc]
            default: updated_desc
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicketSummary"
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total number of tickets matching the filter
  /support/tickets/{id}:
    get:
      operationId: getSupportTicket
      tags: [support-tickets]
      summary: Get ticket detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateSupportTicket
      tags: [support-tickets]
      summary: Update ticket status / priority / assignee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketDetail"
  /support/tickets/{id}/messages:
    post:
      operationId: postSupportTicketMessage
      tags: [support-tickets]
      summary: Add a public reply or internal note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketMessageInput"
      responses:
        "201":
          description: Message added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketDetail"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    TicketStatus:
      type: string
      enum: [open, pending, resolved]
    TicketPriority:
      type: string
      enum: [low, normal, high, urgent]
    RequesterRole:
      type: string
      enum: [rider, driver]
...
    SupportRequester:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/RequesterRole"
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    SupportRelatedTrip:
      type: object
      required: [id, origin, destination, departureAt]
      properties:
        id:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureAt:
          type: string
          format: date-time
        driverName:
          type: ["string", "null"]
        priceCents:
          type: ["integer", "null"]
    SupportTicketSummary:
      type: object
      required:
        - id
        - subject
        - status
        - priority
        - requester
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        subject:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        requester:
          $ref: "#/components/schemas/SupportRequester"
        tripId:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        assigneeName:
          type: ["string", "null"]
        lastMessagePreview:
          type: ["string", "null"]
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SupportTicketMessage:
      type: object
      required: [id, body, authorType, authorName, internal, createdAt]
      properties:
        id:
          type: string
        body:
          type: string
        authorType:
          type: string
          enum: [customer, agent, system]
        authorName:
          type: string
        internal:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SupportTicketDetail:
      type: object
      required:
        - id
        - subject
        - status
        - priority
        - requester
        - messages
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        subject:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        requester:
          $ref: "#/components/schemas/SupportRequester"
        trip:
          oneOf:
            - $ref: "#/components/schemas/SupportRelatedTrip"
            - type: "null"
        assigneeId:
          type: ["string", "null"]
        assigneeName:
          type: ["string", "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportTicketMessage"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SupportTicketUpdate:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        assigneeId:
          type: ["string", "null"]
    SupportTicketMessageInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
        internal:
          type: boolean
...